Legger til behandlingsresultat på eksisterende tester slik at de kjører. Renaming i beregning-testmappen slik at navngiving i enhetstester og integrasjonstester er konsistent
