+++ src/test/kotlin/no/nav/familie/ba/sak/behandling/BehandlingIntegrationTest.kt
@@ -2,10 +2,7 @@ package no.nav.familie.ba.sak.behandling
 
 import com.github.tomakehurst.wiremock.client.WireMock.*
 import io.mockk.*
-import no.nav.familie.ba.sak.behandling.domene.BehandlingKategori
-import no.nav.familie.ba.sak.behandling.domene.BehandlingRepository
-import no.nav.familie.ba.sak.behandling.domene.BehandlingType
-import no.nav.familie.ba.sak.behandling.domene.BehandlingUnderkategori
+import no.nav.familie.ba.sak.behandling.domene.*
 import no.nav.familie.ba.sak.behandling.fagsak.FagsakRequest
 import no.nav.familie.ba.sak.behandling.fagsak.FagsakService
 import no.nav.familie.ba.sak.behandling.grunnlag.personopplysninger.PersongrunnlagService
@@ -13,6 +10,7 @@ import no.nav.familie.ba.sak.behandling.steg.StegType
 import no.nav.familie.ba.sak.behandling.vedtak.VedtakRepository
 import no.nav.familie.ba.sak.behandling.vedtak.VedtakService
 import no.nav.familie.ba.sak.behandling.vedtak.Ytelsetype
+import no.nav.familie.ba.sak.behandling.vilkår.PeriodeResultat
 import no.nav.familie.ba.sak.beregning.*
 import no.nav.familie.ba.sak.common.*
 import no.nav.familie.ba.sak.integrasjoner.domene.Personinfo
@@ -22,6 +20,7 @@ import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.familie.kontrakter.felles.objectMapper
 import no.nav.familie.prosessering.domene.Task
 import no.nav.familie.prosessering.domene.TaskRepository
+import no.nav.nare.core.evaluations.Resultat
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Tag
@@ -59,6 +58,12 @@ class BehandlingIntegrationTest {
 
     @Autowired
     lateinit var beregningService: BeregningService
+
+    @Autowired
+    lateinit var behandlingResultatRepository: BehandlingResultatRepository
+
+    @Autowired
+    lateinit var behandlingResultatService: BehandlingResultatService
 
     @Autowired
     lateinit var fagsakService: FagsakService
@@ -146,8 +151,6 @@ class BehandlingIntegrationTest {
     @Test
      fun `Bruk samme behandling hvis nytt barn kommer på fagsak med aktiv behandling`() {
         val morId = randomFnr()
-        val barnId = randomFnr()
-        val barn2Id = randomFnr()
 
         val fagsak = fagsakService.hentEllerOpprettFagsakForPersonIdent(morId)
         behandlingService.opprettBehandling(nyOrdinærBehandling(morId))
@@ -171,6 +174,8 @@ class BehandlingIntegrationTest {
         val søkerFnr = randomFnr()
         val barn1Fnr = randomFnr()
         val barn2Fnr = randomFnr()
+        val stønadFom = LocalDate.of(2020, 1, 1)
+        val stønadTom = stønadFom.plusYears(17)
 
         fagsakService.hentEllerOpprettFagsak(FagsakRequest(personIdent = søkerFnr))
         val behandling = behandlingService.opprettBehandling(nyOrdinærBehandling(søkerFnr))
@@ -178,6 +183,10 @@ class BehandlingIntegrationTest {
         val personopplysningGrunnlag =
                 lagTestPersonopplysningGrunnlag(behandling.id, søkerFnr, listOf(barn1Fnr, barn2Fnr))
         persongrunnlagService.lagreOgDeaktiverGammel(personopplysningGrunnlag)
+
+        val behandlingResultat = BehandlingResultat(behandling = behandling)
+        behandlingResultat.periodeResultater = lagPeriodeResultaterForSøkerOgToBarn(behandlingResultat, søkerFnr, barn1Fnr, barn2Fnr, stønadFom, stønadTom)
+        behandlingResultatRepository.save(behandlingResultat)
 
         val nyBeregning = NyBeregning(listOf(
                 PersonBeregning(barn1Fnr,
@@ -238,6 +247,7 @@ class BehandlingIntegrationTest {
 
         val dato_2020_01_01 = LocalDate.of(2020, 1, 1)
         val dato_2020_10_01 = LocalDate.of(2020, 10, 1)
+        val stønadTom = dato_2020_01_01.plusYears(17)
 
         fagsakService.hentEllerOpprettFagsak(FagsakRequest(personIdent = søkerFnr))
         val behandling = behandlingService.opprettBehandling(nyOrdinærBehandling(søkerFnr))
@@ -250,6 +260,32 @@ class BehandlingIntegrationTest {
                 behandling = behandling,
                 personopplysningGrunnlag = personopplysningGrunnlag,
                 ansvarligSaksbehandler = "saksbehandler1")
+
+        val behandlingResultat = BehandlingResultat(behandling = behandling)
+        behandlingResultat.periodeResultater = setOf(
+                lagPeriodeResultat(
+                        søkerFnr,
+                        behandlingResultat = behandlingResultat,
+                        resultat = Resultat.JA,
+                        periodeFom = dato_2020_01_01.minusMonths(1),
+                        periodeTom = stønadTom
+                ),
+                lagPeriodeResultat(
+                        barn1Fnr,
+                        behandlingResultat = behandlingResultat,
+                        resultat = Resultat.JA,
+                        periodeFom = dato_2020_01_01.minusMonths(1),
+                        periodeTom = stønadTom
+                ),
+                lagPeriodeResultat(
+                        barn2Fnr,
+                        behandlingResultat = behandlingResultat,
+                        resultat = Resultat.JA,
+                        periodeFom = dato_2020_10_01.minusMonths(1),
+                        periodeTom = stønadTom
+                )
+        )
+        behandlingResultatRepository.save(behandlingResultat)
 
         val personBeregninger = listOf(
                 PersonBeregning(barn1Fnr, 1054, dato_2020_01_01, Ytelsetype.ORDINÆR_BARNETRYGD),
@@ -270,7 +306,7 @@ class BehandlingIntegrationTest {
         Assertions.assertTrue(dato_2020_01_01 < restVedtakBarnMap[barn1Fnr]!!.stønadTom)
         Assertions.assertEquals(Ytelsetype.ORDINÆR_BARNETRYGD,restVedtakBarnMap[barn1Fnr]!!.type)
 
-        Assertions.assertEquals(1354,restVedtakBarnMap[barn2Fnr]!!.beløp)
+        Assertions.assertEquals(1054, restVedtakBarnMap[barn2Fnr]!!.beløp)
         Assertions.assertEquals(dato_2020_10_01, restVedtakBarnMap[barn2Fnr]!!.stønadFom)
         Assertions.assertTrue(dato_2020_10_01 < restVedtakBarnMap[barn2Fnr]!!.stønadTom)
         Assertions.assertEquals(Ytelsetype.ORDINÆR_BARNETRYGD,restVedtakBarnMap[barn2Fnr]!!.type)
@@ -288,6 +324,7 @@ class BehandlingIntegrationTest {
         val dato_2020_10_01 = LocalDate.of(2020, 10, 1)
         val dato_2021_01_01 = LocalDate.of(2021, 1, 1)
         val dato_2021_10_01 = LocalDate.of(2021, 10, 1)
+        val stønadTom = dato_2020_01_01.plusYears(17)
 
         fagsakService.hentEllerOpprettFagsak(FagsakRequest(personIdent = søkerFnr))
         val behandling = behandlingService.opprettBehandling(nyOrdinærBehandling(søkerFnr))
@@ -303,12 +340,21 @@ class BehandlingIntegrationTest {
                 personopplysningGrunnlag = personopplysningGrunnlag,
                 ansvarligSaksbehandler = "saksbehandler1")
 
+        val behandlingResultat1 = BehandlingResultat(behandling = behandling)
+        behandlingResultat1.periodeResultater = lagPeriodeResultaterForSøkerOgToBarn(behandlingResultat1, søkerFnr, barn1Fnr, barn2Fnr, dato_2020_01_01.minusMonths(1), stønadTom)
+        behandlingResultatRepository.save(behandlingResultat1)
+
         val førsteBeregning = NyBeregning(listOf(
                 PersonBeregning(barn1Fnr, 1054, dato_2020_01_01, Ytelsetype.ORDINÆR_BARNETRYGD),
                 PersonBeregning(barn2Fnr, 1354, dato_2020_10_01, Ytelsetype.ORDINÆR_BARNETRYGD)
         ))
 
         beregningService.oppdaterBehandlingMedBeregning(behandling, personopplysningGrunnlag, førsteBeregning)
+
+
+        val behandlingResultat2 = BehandlingResultat(behandling = behandling)
+        behandlingResultat2.periodeResultater = lagPeriodeResultaterForSøkerOgToBarn(behandlingResultat2, søkerFnr, barn1Fnr, barn3Fnr, dato_2021_01_01.minusMonths(1), stønadTom)
+        behandlingResultatService.lagreNyOgDeaktiverGammel(behandlingResultat2)
 
         val andreBeregning = NyBeregning(listOf(
                 PersonBeregning(barn1Fnr, 970, dato_2021_01_01, Ytelsetype.MANUELL_VURDERING),
@@ -322,14 +368,14 @@ class BehandlingIntegrationTest {
                 .associateBy({it.barn}, {it.ytelsePerioder[0]} )
 
         Assertions.assertEquals(2, restVedtakBarnMap.size)
-        Assertions.assertEquals(970,restVedtakBarnMap[barn1Fnr]!!.beløp)
+        Assertions.assertEquals(1054,restVedtakBarnMap[barn1Fnr]!!.beløp)
         Assertions.assertEquals(dato_2021_01_01, restVedtakBarnMap[barn1Fnr]!!.stønadFom)
         Assertions.assertTrue(dato_2021_01_01 < restVedtakBarnMap[barn1Fnr]!!.stønadTom)
-        Assertions.assertEquals(Ytelsetype.MANUELL_VURDERING,restVedtakBarnMap[barn1Fnr]!!.type)
 
-        Assertions.assertEquals(314,restVedtakBarnMap[barn3Fnr]!!.beløp)
-        Assertions.assertEquals(dato_2021_10_01, restVedtakBarnMap[barn3Fnr]!!.stønadFom)
-        Assertions.assertTrue(dato_2021_10_01 < restVedtakBarnMap[barn3Fnr]!!.stønadTom)
-        Assertions.assertEquals(Ytelsetype.EØS,restVedtakBarnMap[barn3Fnr]!!.type)
+        Assertions.assertEquals(1054,restVedtakBarnMap[barn3Fnr]!!.beløp)
+        Assertions.assertEquals(dato_2021_01_01, restVedtakBarnMap[barn3Fnr]!!.stønadFom)
+        Assertions.assertTrue(dato_2021_01_01 < restVedtakBarnMap[barn3Fnr]!!.stønadTom)
+
+        Assertions.assertNull(restVedtakBarnMap[barn2Fnr])
     }
 }
