+++ src/test/kotlin/no/nav/familie/ba/sak/økonomi/ØkonomiIntegrasjonTest.kt
@@ -2,6 +2,8 @@ package no.nav.familie.ba.sak.økonomi
 
 import com.github.tomakehurst.wiremock.client.WireMock.*
 import no.nav.familie.ba.sak.behandling.BehandlingService
+import no.nav.familie.ba.sak.behandling.domene.Behandling
+import no.nav.familie.ba.sak.behandling.domene.BehandlingResultat
 import no.nav.familie.ba.sak.behandling.domene.BehandlingResultatService
 import no.nav.familie.ba.sak.behandling.domene.BehandlingStatus
 import no.nav.familie.ba.sak.behandling.fagsak.FagsakService
@@ -13,10 +15,8 @@ import no.nav.familie.ba.sak.behandling.vedtak.Ytelsetype
 import no.nav.familie.ba.sak.beregning.BeregningService
 import no.nav.familie.ba.sak.beregning.NyBeregning
 import no.nav.familie.ba.sak.beregning.PersonBeregning
-import no.nav.familie.ba.sak.common.lagBehandling
-import no.nav.familie.ba.sak.common.lagBehandlingResultat
-import no.nav.familie.ba.sak.common.lagTestPersonopplysningGrunnlag
-import no.nav.familie.ba.sak.common.randomFnr
+import no.nav.familie.ba.sak.beregning.lagPeriodeResultat
+import no.nav.familie.ba.sak.common.*
 import no.nav.familie.ba.sak.config.ApplicationConfig
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.familie.kontrakter.felles.objectMapper
@@ -26,17 +26,22 @@ import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle
+import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock
 import org.springframework.test.context.ActiveProfiles
+import org.springframework.test.context.ContextConfiguration
+import org.springframework.test.context.junit.jupiter.SpringExtension
 import java.time.LocalDate
 
 @SpringBootTest(classes = [ApplicationConfig::class],
                 properties = ["FAMILIE_OPPDRAG_API_URL=http://localhost:28085/api",
                     "FAMILIE_BA_DOKGEN_API_URL=http://localhost:28085/api",
                     "FAMILIE_INTEGRASJONER_API_URL=http://localhost:28085/api"])
-@ActiveProfiles("dev", "mock-oauth")
+@ExtendWith(SpringExtension::class)
+@ContextConfiguration(initializers = [DbContainerInitializer::class])
+@ActiveProfiles("postgres", "mock-dokgen", "mock-oauth")
 @TestInstance(Lifecycle.PER_CLASS)
 @AutoConfigureWireMock(port = 28085)
 class ØkonomiIntegrasjonTest {
@@ -67,6 +72,8 @@ class ØkonomiIntegrasjonTest {
     fun `Iverksett vedtak på aktiv behandling`() {
         val fnr = randomFnr()
         val barnFnr = randomFnr()
+        val stønadFom = LocalDate.now()
+        val stønadTom = stønadFom.plusYears(17)
 
         val responseBody = Ressurs.Companion.success("ok")
         stubFor(get(urlEqualTo("/api/aktoer/v1"))
@@ -84,7 +91,7 @@ class ØkonomiIntegrasjonTest {
         val fagsak = fagsakService.hentEllerOpprettFagsakForPersonIdent(fnr)
         var behandling = behandlingService.lagreNyOgDeaktiverGammelBehandling(lagBehandling(fagsak))
 
-        val behandlingResultat = lagBehandlingResultat(fnr, behandling, Resultat.JA)
+        val behandlingResultat = lagBehandlingResultat(behandling, fnr, barnFnr, stønadFom, stønadTom)
 
         behandlingResultatService.lagreNyOgDeaktiverGammel(behandlingResultat)
         Assertions.assertNotNull(behandling.fagsak.id)
@@ -127,6 +134,8 @@ class ØkonomiIntegrasjonTest {
     fun `Hent behandlinger for løpende fagsaker til konsistensavstemming mot økonomi`() {
         val fnr = randomFnr()
         val barnFnr = randomFnr()
+        val stønadFom = LocalDate.now()
+        val stønadTom = stønadFom.plusYears(17)
 
         stubFor(post(anyUrl())
                         .willReturn(aResponse()
@@ -146,6 +155,9 @@ class ØkonomiIntegrasjonTest {
                 lagTestPersonopplysningGrunnlag(behandling.id, fnr, listOf(barnFnr))
         personopplysningGrunnlagRepository.save(personopplysningGrunnlag)
         vedtakService.lagreOgDeaktiverGammel(vedtak)
+
+        val behandlingResultat = lagBehandlingResultat(behandling, fnr, barnFnr, stønadFom, stønadTom)
+        behandlingResultatService.lagreNyOgDeaktiverGammel(behandlingResultat)
 
         val nyBeregning = NyBeregning(
                 listOf(PersonBeregning(ident = barnFnr,
@@ -168,5 +180,26 @@ class ØkonomiIntegrasjonTest {
         val oppdragIdListe = behandlingService.hentGjeldendeBehandlingerForLøpendeFagsaker()
 
         Assertions.assertTrue(oppdragIdListe.contains(OppdragId(fnr, behandling.id)))
+    }
+
+    private fun lagBehandlingResultat(behandling: Behandling, søkerFnr: String, barnFnr: String, stønadFom: LocalDate, stønadTom: LocalDate): BehandlingResultat {
+        val behandlingResultat = BehandlingResultat(behandling = behandling)
+        behandlingResultat.periodeResultater = setOf(
+                lagPeriodeResultat(
+                        søkerFnr,
+                        behandlingResultat = behandlingResultat,
+                        resultat = Resultat.JA,
+                        periodeFom = stønadFom,
+                        periodeTom = stønadTom
+                ),
+                lagPeriodeResultat(
+                        barnFnr,
+                        behandlingResultat = behandlingResultat,
+                        resultat = Resultat.JA,
+                        periodeFom = stønadFom,
+                        periodeTom = stønadTom
+                )
+        )
+        return behandlingResultat
     }
 }
